# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ConceptOut": {
        "properties": {
          "aliases": {
            "items": {
              "type": "string"
            },
            "title": "Aliases",
            "type": "array"
          },
          "bloom_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bloom Level"
          },
          "canonical_key": {
            "title": "Canonical Key",
            "type": "string"
          },
          "definition": {
            "title": "Definition",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "importance": {
            "title": "Importance",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "canonical_key",
          "name",
          "definition",
          "aliases",
          "importance",
          "bloom_level"
        ],
        "title": "ConceptOut",
        "type": "object"
      },
...
      "EdgeOut": {
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "dependent_concept_id": {
            "format": "uuid",
            "title": "Dependent Concept Id",
            "type": "string"
          },
          "prerequisite_concept_id": {
            "format": "uuid",
            "title": "Prerequisite Concept Id",
            "type": "string"
          },
          "strength": {
            "title": "Strength",
            "type": "string"
          }
        },
        "required": [
          "prerequisite_concept_id",
          "dependent_concept_id",
          "strength",
          "confidence"
        ],
        "title": "EdgeOut",
        "type": "object"
      },
...
      "GraphOut": {
        "properties": {
          "concepts": {
            "items": {
              "$ref": "#/components/schemas/ConceptOut"
            },
            "title": "Concepts",
            "type": "array"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/EdgeOut"
            },
            "title": "Edges",
            "type": "array"
          }
        },
        "required": [
          "concepts",
          "edges"
        ],
        "title": "GraphOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LessonConceptOut": {
        "properties": {
          "concept_id": {
            "format": "uuid",
            "title": "Concept Id",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "weight": {
            "title": "Weight",
            "type": "number"
          }
        },
        "required": [
          "concept_id",
          "role",
          "weight"
        ],
        "title": "LessonConceptOut",
        "type": "object"
      },
...
      "LessonOut": {
        "properties": {
          "concepts": {
            "items": {
              "$ref": "#/components/schemas/LessonConceptOut"
            },
            "title": "Concepts",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "objective": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Objective"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "objective",
          "concepts"
        ],
        "title": "LessonOut",
        "type": "object"
      },
      "LessonRename": {
        "properties": {
          "lesson_id": {
            "format": "uuid",
            "title": "Lesson Id",
            "type": "string"
          },
          "title": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "lesson_id",
          "title"
        ],
        "title": "LessonRename",
        "type": "object"
      },
...
      "ModuleOut": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "lessons": {
            "items": {
              "$ref": "#/components/schemas/LessonOut"
            },
            "title": "Lessons",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "lessons"
        ],
        "title": "ModuleOut",
        "type": "object"
      },
...
      "PublishedStructureOut": {
        "properties": {
          "active_version_id": {
            "format": "uuid",
            "title": "Active Version Id",
            "type": "string"
          },
          "course_id": {
            "format": "uuid",
            "title": "Course Id",
            "type": "string"
          }
        },
        "required": [
          "course_id",
          "active_version_id"
        ],
        "title": "PublishedStructureOut",
        "type": "object"
      },
...
      "StructureOut": {
        "properties": {
          "concept_carryover_map": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Concept Carryover Map"
          },
          "modules": {
            "items": {
              "$ref": "#/components/schemas/ModuleOut"
            },
            "title": "Modules",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "validation_errors": {
            "items": {
              "type": "string"
            },
            "title": "Validation Errors",
            "type": "array"
          },
          "version_id": {
            "format": "uuid",
            "title": "Version Id",
            "type": "string"
          },
          "version_number": {
            "title": "Version Number",
            "type": "integer"
          }
        },
        "required": [
          "version_id",
          "version_number",
          "status",
          "validation_errors",
          "concept_carryover_map",
          "modules"
        ],
        "title": "StructureOut",
        "type": "object"
      },
      "StructureUpdateIn": {
        "description": "Minimal edit surface this phase supports: renaming lessons.",
        "properties": {
          "lesson_renames": {
            "items": {
              "$ref": "#/components/schemas/LessonRename"
            },
            "title": "Lesson Renames",
            "type": "array"
          }
        },
        "title": "StructureUpdateIn",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Backend Service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/courses/{course_id}/graph": {
      "get": {
        "operationId": "get_graph_api_v1_courses__course_id__graph_get",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version Id"
            }
          },
          {
            "in": "header",
            "name": "x-user-email",
            "required": true,
            "schema": {
              "title": "X-User-Email",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": true,
            "schema": {
              "title": "X-Internal-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Graph",
        "tags": [
          "curriculum"
        ]
      }
    },
...
    "/api/v1/courses/{course_id}/publish-structure": {
      "post": {
        "description": "Explicit confirmation, per the mandate: this is the only route that ever\nchanges course.active_version_id. Defaults to publishing the latest\ngenerated version (the one the review gate showed).",
        "operationId": "publish_structure_api_v1_courses__course_id__publish_structure_post",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version Id"
            }
          },
          {
            "in": "header",
            "name": "x-user-email",
            "required": true,
            "schema": {
              "title": "X-User-Email",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": true,
            "schema": {
              "title": "X-Internal-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishedStructureOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Publish Structure",
        "tags": [
          "curriculum"
        ]
      }
    },
...
    "/api/v1/courses/{course_id}/structure": {
      "get": {
        "description": "The outline review gate's read side: the latest generated version,\nwhatever its status, so a learner can review a draft before publishing.",
        "operationId": "get_structure_api_v1_courses__course_id__structure_get",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-email",
            "required": true,
            "schema": {
              "title": "X-User-Email",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": true,
            "schema": {
              "title": "X-Internal-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StructureOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Structure",
        "tags": [
          "curriculum"
        ]
      },
      "put": {
        "description": "Cheap correction before expensive [re]generation: edit the reviewed\ndraft's lesson titles. Reordering modules and dropping concepts are not\nimplemented yet.",
        "operationId": "update_structure_api_v1_courses__course_id__structure_put",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-email",
            "required": true,
            "schema": {
              "title": "X-User-Email",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": true,
            "schema": {
              "title": "X-Internal-Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StructureUpdateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StructureOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Structure",
        "tags": [
          "curriculum"
        ]
      }
    },
...
  }
}